assignment3--some changes

+++ src/editor/terminal.rs
@@ -9,14 +9,20 @@ use std::io::Write;
 
 pub struct Terminal; //无字段的结构体
 pub struct Size {
-    pub height: u16,
-    pub width: u16,
+    pub height: usize,
+    pub width: usize,
 }
 #[derive(Copy, Clone)]
 pub struct Position {
-    pub x: u16,
-    pub y: u16,
+    pub x: usize,
+    pub y: usize,
 }
+/// Represents the Terminal.
+/// Edge Case for platforms where `usize` < `u16`:
+/// Regardless of the actual size of the Terminal, this representation
+/// only spans over at most `usize::MAX` or `u16::size` rows/columns, whichever is smaller.
+/// Each size returned truncates to min(`usize::MAX`, `u16::MAX`)
+/// And should you attempt to set the cursor out of these bounds, it will also be truncated.
 impl Terminal {
     pub fn initialize() -> Result<(), Error> {
         enable_raw_mode()?;
@@ -38,14 +44,22 @@ impl Terminal {
         Self::queue_command(Clear(ClearType::CurrentLine))?;
         Ok(())
     }
+    ///Moves the cursor to eht given Position
+    /// #Arguments
+    /// * 'Position' - the 'Position' to move the cursor to.Will be truncated to 'u16::MAX' if bigger
     pub fn move_cursor_to(position: Position) -> Result<(), Error> {
-        Self::queue_command(MoveTo(position.x, position.y))?;
+        #[allow(clippy::as_conversions, clippy::cast_possible_truncation)]
+        Self::queue_command(MoveTo(position.x as u16, position.y as u16))?;
         Ok(())
     }
+    ///Returns the current size of the terminal
+    /// Edge Case for systems with `usize` < `u16`:
+    /// * A `Size` representing the terminal size. Any coordinate `z` truncated to `usize` if `usize` < `z` < `u16`
     pub fn size() -> Result<Size, Error> {
+        #[allow(clippy::as_conversions)]
         Ok(Size {
-            height: size()?.1,
-            width: size()?.0,
+            height: size()?.1 as usize,
+            width: size()?.0 as usize,
         })
     }
     pub fn hide_cursor() -> Result<(), Error> {
